feat: rename auth -> oauth (path, controller)
