+++ website/src/template/oauth.html
@@ -1,0 +1,3 @@
+<div class="oauth">
+    {provider} 認證中
+</div>
